+++ cli/test-driver/src/main.rs
@@ -226,7 +226,6 @@ impl BackendTestContext {
     async fn run_verification(&self, job: JobKind) -> Result<()> {
         let dir = self.path_to_snapshots().await?;
         let output = match self.backend {
-            BackendName::Fstar => run_fstar(true, dir.clone()).await?,
             BackendName::LegacyLean => run_lean(dir.clone()).await?,
             _ => unreachable!(),
         };